% Generated by roxygen2 (4.0.1): do not edit by hand
\name{orphan}
\alias{orphan}
\title{Find Orphan Values}
\usage{
orphan(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Finds values that occur exactly once in a vector.
}
\details{
Returns the unique values in the same order that they would be
  returned in a call to \code{\link[base]{unique}}.
}
\examples{
orphan(c(NA, 1:3, 3, 4:6, 3, NA, 4))
}
\seealso{
\code{\link[base]{unique}}
}
\keyword{orphan,}
\keyword{unique}

