% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{tuple-package}
\alias{tuple-package}
\title{The tuple Package}
\description{
Find every match, or orphan, duplicate or triplicate values.
}
\details{
\tabular{ll}{
Package: \tab tuple\cr
Type: \tab Package\cr
Version: \tab 0.3-06\cr
Date: \tab 2014-07-14\cr
Depends: \tab R (>= 2.10.0)\cr
Encoding: \tab UTF-8\cr
License: \tab LGPL-3\cr
LazyLoad: \tab no\cr
URL: \tab http://statistics.lazaridis.eu\cr
}

Functions to find all matches, or the orphan, duplicate or triplicate values in a vector.
}
\author{
Emmanuel Lazaridis
}
\keyword{package}

